indclass {indclass}	R Documentation

indclass: a collection of tools for working with industry codes

Description

indclass is supposed to make working with industry codes simpler and faster.

[Package indclass version 0.1.0 Index]